Refactor context property handling; add Event.profile property; move mothballed v1p1-extension tests to v1p2.
