+++ src/main/java/org/imsglobal/caliper/config/Config.java
@@ -26,7 +26,6 @@ import org.imsglobal.caliper.validators.SensorValidator;
 public class Config {
     private final DataFormat dataFormat;
     private final String dataVersion;
-    private final String jsonldExternalCaliperContext;
     private final int uuidVersion;
 
     /**
@@ -57,7 +56,6 @@ public class Config {
 
         this.dataFormat = (builder.dataFormat != null) ? builder.dataFormat : DATA_FORMAT;
         this.dataVersion = SensorValidator.chkStrValue(builder.dataVersion, DATA_VERSION);
-        this.jsonldExternalCaliperContext = SensorValidator.chkStrValue(builder.jsonldExternalCaliperContext, JSONLD_EXTERNAL_CALIPER_CONTEXT);
         this.uuidVersion = SensorValidator.chkIntValue(builder.uuidVersion, UUID_VERSION);
     }
 
@@ -78,14 +76,6 @@ public class Config {
     }
 
     /**
-     * Get the external Caliper JSON-LD context URI.
-     * @return the external Caliper JSON-LD context URI
-     */
-    public String getJsonldExternalCaliperContext() {
-        return jsonldExternalCaliperContext;
-    }
-
-    /**
      * Get the UUID version to be used when minting Event UUIDs.
      * @return UUID version
      */
@@ -99,7 +89,6 @@ public class Config {
     public static class ConfigBuilder {
         private DataFormat dataFormat;
         private String dataVersion;
-        private String jsonldExternalCaliperContext;
         private int uuidVersion = 4;
 
         /**
@@ -124,15 +113,6 @@ public class Config {
          */
         public ConfigBuilder dataVersion(final String dataVersion) {
             this.dataVersion = dataVersion;
-            return this;
-        }
-
-        /**
-         * @param jsonldExternalCaliperContext
-         * @return builder
-         */
-        public ConfigBuilder jsonldExternalCaliperContext(final String jsonldExternalCaliperContext) {
-            this.jsonldExternalCaliperContext = jsonldExternalCaliperContext;
             return this;
         }
