+++ src/main/java/org/imsglobal/caliper/context/JsonldStringContext.java
@@ -65,12 +65,4 @@ public class JsonldStringContext implements JsonldContext, CaliperCoercible {
     public static JsonldStringContext create(String id) {
         return new JsonldStringContext(id);
     }
-
-    /**
-     * Factory method that returns the default IMS Caliper external context IRI.
-     * @return
-     */
-    public static JsonldStringContext getDefault() {
-        return new JsonldStringContext(Config.JSONLD_EXTERNAL_CALIPER_CONTEXT);
-    }
 }
