+++ src/test/java/org/imsglobal/caliper/v1p1/entities/CourseOfferingTest.java
@@ -20,6 +20,7 @@ package org.imsglobal.caliper.v1p1.entities;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseOffering;
 import org.joda.time.DateTime;
@@ -43,7 +44,7 @@ public class CourseOfferingTest {
     public void setUp() throws Exception {
 
         entity = CourseOffering.builder()
-                .context(JsonldStringContext.getDefault())
+                .context(JsonldStringContext.create(CaliperJsonldContext.V1P1.value()))
                 .id(BASE_IRI.concat("/terms/201601/courses/7"))
                 .courseNumber("CPS 435")
                 .academicSession("Fall 2016")
