+++ src/test/java/org/imsglobal/caliper/v1p1/entities/GroupTest.java
@@ -21,6 +21,7 @@ package org.imsglobal.caliper.v1p1.entities;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Lists;
 import org.imsglobal.caliper.TestUtils;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CaliperAgent;
 import org.imsglobal.caliper.entities.agent.Person;
@@ -77,7 +78,7 @@ public class GroupTest {
             .build();
 
         entity = Group.builder()
-            .context(JsonldStringContext.getDefault())
+            .context(JsonldStringContext.create(CaliperJsonldContext.V1P1.value()))
             .id(BASE_IRI.concat("/terms/201601/courses/7/sections/1/groups/2"))
             .name("Discussion Group 2")
             .subOrganizationOf(courseSection)
