+++ src/test/java/org/imsglobal/caliper/v1p1/entities/PersonTest.java
@@ -20,6 +20,7 @@ package org.imsglobal.caliper.v1p1.entities;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.Person;
 import org.joda.time.DateTime;
@@ -43,7 +44,7 @@ public class PersonTest {
     public void setUp() throws Exception {
 
         entity = Person.builder()
-            .context(JsonldStringContext.getDefault())
+            .context(JsonldStringContext.create(CaliperJsonldContext.V1P1.value()))
             .id(BASE_IRI.concat("/users/554433"))
             .dateCreated(new DateTime(2016, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
             .dateModified(new DateTime(2016, 9, 2, 11, 30, 0, 0, DateTimeZone.UTC))
