+++ src/test/java/org/imsglobal/caliper/v1p1/events/AnnotationEventBookmarkedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -63,7 +65,7 @@ public class AnnotationEventBookmarkedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:d4618c23-d612-4709-8d9a-478d87808067";
 
@@ -140,7 +142,7 @@ public class AnnotationEventBookmarkedTest {
      * @param action
      * @return event
      */
-    private AnnotationEvent buildEvent(Action action) {
+    private AnnotationEvent buildEvent(CaliperAction action) {
         return AnnotationEvent.builder()
             .context(context)
             .id(id)
