+++ src/test/java/org/imsglobal/caliper/v1p1/events/AnnotationEventTaggedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -67,7 +69,7 @@ public class AnnotationEventTaggedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:b2009c63-2659-4cd2-b71e-6e03c498f02b";
 
@@ -148,7 +150,7 @@ public class AnnotationEventTaggedTest {
      * @param action
      * @return event
      */
-    private AnnotationEvent buildEvent(Action action) {
+    private AnnotationEvent buildEvent(CaliperAction action) {
         return AnnotationEvent.builder()
             .context(context)
             .id(id)
