+++ src/test/java/org/imsglobal/caliper/v1p1/events/AssessmentEventStartedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -62,7 +64,7 @@ public class AssessmentEventStartedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:27734504-068d-4596-861c-2315be33a2a2";
 
@@ -138,7 +140,7 @@ public class AssessmentEventStartedTest {
      * @param action
      * @return event
      */
-    private AssessmentEvent buildEvent(Action action) {
+    private AssessmentEvent buildEvent(CaliperAction action) {
         return AssessmentEvent.builder()
             .context(context)
             .id(id)
