+++ src/test/java/org/imsglobal/caliper/v1p1/events/AssessmentEventSubmittedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -62,7 +64,7 @@ public class AssessmentEventSubmittedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:dad88464-0c20-4a19-a1ba-ddf2f9c3ff33";
 
@@ -139,7 +141,7 @@ public class AssessmentEventSubmittedTest {
      * @param action
      * @return event
      */
-    private AssessmentEvent buildEvent(Action action) {
+    private AssessmentEvent buildEvent(CaliperAction action) {
         return AssessmentEvent.builder()
             .context(context)
             .id(id)
