+++ src/test/java/org/imsglobal/caliper/v1p1/events/AssessmentItemEventStartedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -63,7 +65,7 @@ public class AssessmentItemEventStartedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:1b557176-ba67-4624-b060-6bee670a3d8e";
 
@@ -144,7 +146,7 @@ public class AssessmentItemEventStartedTest {
      * @param action
      * @return event
      */
-    private AssessmentItemEvent buildEvent(Action action) {
+    private AssessmentItemEvent buildEvent(CaliperAction action) {
         return AssessmentItemEvent.builder()
             .context(context)
             .id(id)
