+++ src/test/java/org/imsglobal/caliper/v1p1/events/BasicEventModifiedExtendedTest.java
@@ -22,6 +22,8 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.Person;
@@ -55,7 +57,7 @@ public class BasicEventModifiedExtendedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:5973dcd9-3126-4dcc-8fd8-8153a155361c";
 
@@ -109,7 +111,7 @@ public class BasicEventModifiedExtendedTest {
      * @param action
      * @return event
      */
-    private Event buildEvent(Action action) {
+    private Event buildEvent(CaliperAction action) {
         return Event.builder()
             .context(context)
             .id(id)
