+++ src/test/java/org/imsglobal/caliper/v1p1/events/ForumEventSubscribedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -60,7 +62,7 @@ public class ForumEventSubscribedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:a2f41f9c-d57d-4400-b3fe-716b9026334e";
 
@@ -123,7 +125,7 @@ public class ForumEventSubscribedTest {
      * @param action
      * @return event
      */
-    private ForumEvent buildEvent(Action action) {
+    private ForumEvent buildEvent(CaliperAction action) {
         return ForumEvent.builder()
             .context(context)
             .id(id)
