+++ src/test/java/org/imsglobal/caliper/v1p1/events/GradeEventGradedItemTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -58,7 +60,7 @@ public class GradeEventGradedItemTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:12c05c4e-253f-4073-9f29-5786f3ff3f36";
 
@@ -137,7 +139,7 @@ public class GradeEventGradedItemTest {
      * @param action
      * @return event
      */
-    private GradeEvent buildEvent(Action action) {
+    private GradeEvent buildEvent(CaliperAction action) {
         return GradeEvent.builder()
             .context(context)
             .id(id)
