+++ src/test/java/org/imsglobal/caliper/v1p1/events/NavigationEventNavigatedToWebPageTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -60,7 +62,7 @@ public class NavigationEventNavigatedToWebPageTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:ff9ec22a-fc59-4ae1-ae8d-2c9463ee2f8f";
 
@@ -125,7 +127,7 @@ public class NavigationEventNavigatedToWebPageTest {
      * @param action
      * @return event
      */
-    private NavigationEvent buildEvent(Action action) {
+    private NavigationEvent buildEvent(CaliperAction action) {
         return NavigationEvent.builder()
             .context(context)
             .id(id)
