+++ src/test/java/org/imsglobal/caliper/v1p1/events/NavigationEventNavigatedToWebPageThinnedTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -59,7 +61,7 @@ public class NavigationEventNavigatedToWebPageThinnedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:71657137-8e6e-44f8-8499-e1c3df6810d2";
 
@@ -117,7 +119,7 @@ public class NavigationEventNavigatedToWebPageThinnedTest {
      * @param action
      * @return event
      */
-    private NavigationEvent buildEvent(Action action) {
+    private NavigationEvent buildEvent(CaliperAction action) {
         return NavigationEvent.builder()
             .context(context)
             .id(id)
