+++ src/test/java/org/imsglobal/caliper/v1p1/events/SessionEventLoggedInExtendedTest.java
@@ -23,6 +23,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Maps;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.Person;
@@ -57,7 +59,7 @@ public class SessionEventLoggedInExtendedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:4ec2c31e-3ec0-4fe1-a017-b81561b075d7";
 
@@ -108,7 +110,7 @@ public class SessionEventLoggedInExtendedTest {
      * @param action
      * @return event
      */
-    private SessionEvent buildEvent(Action action) {
+    private SessionEvent buildEvent(CaliperAction action) {
         return SessionEvent.builder()
             .context(context)
             .id(id)
