+++ src/test/java/org/imsglobal/caliper/v1p1/events/SessionEventLoggedInTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.Person;
@@ -51,7 +53,7 @@ public class SessionEventLoggedInTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:fcd495d0-3740-4298-9bec-1154571dc211";
 
@@ -96,7 +98,7 @@ public class SessionEventLoggedInTest {
      * @param action
      * @return event
      */
-    private SessionEvent buildEvent(Action action) {
+    private SessionEvent buildEvent(CaliperAction action) {
         return SessionEvent.builder()
             .context(context)
             .id(id)
