+++ src/test/java/org/imsglobal/caliper/v1p1/events/ThreadEventMarkedAsReadTest.java
@@ -21,6 +21,8 @@ package org.imsglobal.caliper.v1p1.events;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -61,7 +63,7 @@ public class ThreadEventMarkedAsReadTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:6b20c5ba-301c-4e56-85a0-2f3d9a94c249";
 
@@ -130,7 +132,7 @@ public class ThreadEventMarkedAsReadTest {
      * @param action
      * @return event
      */
-    private ThreadEvent buildEvent(Action action) {
+    private ThreadEvent buildEvent(CaliperAction action) {
         return ThreadEvent.builder()
             .context(context)
             .id(id)
