+++ src/test/java/org/imsglobal/caliper/v1p1/events/ViewEventViewedDocumentExtendedTest.java
@@ -23,6 +23,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Maps;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
+import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.CourseSection;
@@ -64,7 +66,7 @@ public class ViewEventViewedDocumentExtendedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.getDefault();
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P1.value());
 
         id = "urn:uuid:3a9bd869-addc-48b1-80f6-a14b2ff591ed";
 
@@ -131,7 +133,7 @@ public class ViewEventViewedDocumentExtendedTest {
      * @param action
      * @return event
      */
-    private ViewEvent buildEvent(Action action) {
+    private ViewEvent buildEvent(CaliperAction action) {
         return ViewEvent.builder()
             .context(context)
             .id(id)
