+++ src/test/java/org/imsglobal/caliper/v1p2/events/FeedbackEventCommentedTest.java
@@ -16,11 +16,12 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.imsglobal.caliper.v1p1.events;
+package org.imsglobal.caliper.v1p2.events;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
 import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
@@ -28,7 +29,10 @@ import org.imsglobal.caliper.entities.agent.*;
 import org.imsglobal.caliper.entities.resource.DigitalResource;
 import org.imsglobal.caliper.entities.resource.DigitalResourceCollection;
 import org.imsglobal.caliper.entities.session.Session;
-import org.imsglobal.caliper.events.ResourceManagementEvent;
+import org.imsglobal.caliper.entities.survey.Comment;
+import org.imsglobal.caliper.events.FeedbackEvent;
+import org.imsglobal.caliper.profiles.CaliperProfile;
+import org.imsglobal.caliper.profiles.Profile;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.After;
@@ -44,18 +48,18 @@ import java.util.List;
 import static com.yammer.dropwizard.testing.JsonHelpers.jsonFixture;
 
 @Category(org.imsglobal.caliper.UnitTest.class)
-public class ResourceManagementEventCopiedTest {
+public class FeedbackEventCommentedTest {
     private JsonldContext context;
     private String id;
     private Person actor;
     private CourseSection section;
     private DigitalResource object;
-    private DigitalResource generated;
+    private Comment generated;
     private DigitalResourceCollection collection;
     private CourseSection group;
     private List<CaliperAgent> creators;
     private Membership membership;
-    private ResourceManagementEvent event;
+    private FeedbackEvent event;
     private Session session;
     private SoftwareApplication edApp;
 
@@ -64,11 +68,9 @@ public class ResourceManagementEventCopiedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.create(CaliperJsonldContext.V1P1_RESOURCE_MANAGEMENT.value());
-        id = "urn:uuid:d3543a73-e307-4190-a755-5ce7b3187bc5";
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P2.value());
+        id = "urn:uuid:0c81f804-62ee-4953-81c5-62d9579c4369";
         actor = Person.builder().id(BASE_IRI.concat("/users/554433")).build();
-        creators = new ArrayList<CaliperAgent>();
-        creators.add(actor);
         section = CourseSection.builder().id(SECTION_IRI).build();
 
         collection = DigitalResourceCollection.builder()
@@ -80,19 +82,17 @@ public class ResourceManagementEventCopiedTest {
         object = DigitalResource.builder()
             .id(SECTION_IRI.concat("/resources/1/syllabus.pdf"))
             .name("Course Syllabus")
-            .creators(creators)
             .mediaType("application/pdf")
             .isPartOf(collection)
             .dateCreated(new DateTime(2018, 8, 2, 11, 32, 0, 0, DateTimeZone.UTC))
             .build();
 
-        generated = DigitalResource.builder()
-            .id(SECTION_IRI.concat("/resources/1/syllabus_copy.pdf"))
-            .name("Course Syllabus (copy)")
-            .creators(creators)
-            .mediaType("application/pdf")
-            .isPartOf(collection)
-            .dateCreated(new DateTime(2018, 11, 15, 10, 05, 0, 0, DateTimeZone.UTC))
+        generated = Comment.builder()
+            .id(SECTION_IRI.concat("/assess/1/items/6/users/665544/responses/1/comment/1"))
+            .commenter(actor)
+            .commentedOn(object)
+            .value("I like what you did here but you need to improve on...")
+            .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
             .build();
 
         edApp = SoftwareApplication.builder().id(BASE_IRI).coercedToId(true).build();
@@ -108,7 +108,7 @@ public class ResourceManagementEventCopiedTest {
             .member(Person.builder().id(actor.getId()).coercedToId(true).build())
             .organization(CourseSection.builder().id(group.getId()).coercedToId(true).build())
             .status(Status.ACTIVE)
-            .role(Role.INSTRUCTOR)
+            .role(Role.LEARNER)
             .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
             .build();
 
@@ -118,7 +118,7 @@ public class ResourceManagementEventCopiedTest {
             .build();
 
         // Build event
-        event = buildEvent(Action.COPIED);
+        event = buildEvent(Profile.FEEDBACK, Action.COMMENTED);
     }
 
     @Test
@@ -126,7 +126,7 @@ public class ResourceManagementEventCopiedTest {
         ObjectMapper mapper = TestUtils.createCaliperObjectMapper();
         String json = mapper.writeValueAsString(event);
 
-        String fixture = jsonFixture("fixtures/v1p1/caliperEventResourceManagementCopied.json");
+        String fixture = jsonFixture("fixtures/v1p2/caliperEventFeedbackCommented.json");
         JSONAssert.assertEquals(fixture, json, JSONCompareMode.NON_EXTENSIBLE);
     }
 
@@ -140,19 +140,20 @@ public class ResourceManagementEventCopiedTest {
      * @param action
      * @return event
      */
-    private ResourceManagementEvent buildEvent(Action action) {
-        return ResourceManagementEvent.builder()
-                .context(context)
-                .id(id)
-                .actor(actor)
-                .action(action)
-                .object(object)
-                .generated(generated)
-                .eventTime(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
-                .edApp(edApp)
-                .group(group)
-                .membership(membership)
-                .session(session)
-                .build();
+    private FeedbackEvent buildEvent(CaliperProfile profile, CaliperAction action) {
+        return FeedbackEvent.builder()
+            .context(context)
+            .profile(profile)
+            .id(id)
+            .actor(actor)
+            .action(action)
+            .object(object)
+            .generated(generated)
+            .eventTime(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
+            .edApp(edApp)
+            .group(group)
+            .membership(membership)
+            .session(session)
+            .build();
     }
 }
