+++ src/test/java/org/imsglobal/caliper/v1p2/events/ResourceManagementEventCopiedTest.java
@@ -16,24 +16,22 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.imsglobal.caliper.v1p1.events;
+package org.imsglobal.caliper.v1p2.events;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.fest.util.Lists;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
 import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
 import org.imsglobal.caliper.entities.agent.*;
 import org.imsglobal.caliper.entities.resource.DigitalResource;
 import org.imsglobal.caliper.entities.resource.DigitalResourceCollection;
-import org.imsglobal.caliper.entities.scale.LikertScale;
 import org.imsglobal.caliper.entities.session.Session;
-import org.imsglobal.caliper.entities.survey.Comment;
-import org.imsglobal.caliper.entities.survey.Rating;
-import org.imsglobal.caliper.entities.question.RatingScaleQuestion;
-import org.imsglobal.caliper.events.FeedbackEvent;
+import org.imsglobal.caliper.events.ResourceManagementEvent;
+import org.imsglobal.caliper.profiles.CaliperProfile;
+import org.imsglobal.caliper.profiles.Profile;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.After;
@@ -49,21 +47,18 @@ import java.util.List;
 import static com.yammer.dropwizard.testing.JsonHelpers.jsonFixture;
 
 @Category(org.imsglobal.caliper.UnitTest.class)
-public class FeedbackEventRankedTest {
+public class ResourceManagementEventCopiedTest {
     private JsonldContext context;
     private String id;
     private Person actor;
     private CourseSection section;
     private DigitalResource object;
-    private Rating generated;
+    private DigitalResource generated;
     private DigitalResourceCollection collection;
     private CourseSection group;
     private List<CaliperAgent> creators;
     private Membership membership;
-    private RatingScaleQuestion question;
-    private Comment ratingComment;
-    private LikertScale scale;
-    private FeedbackEvent event;
+    private ResourceManagementEvent event;
     private Session session;
     private SoftwareApplication edApp;
 
@@ -72,8 +67,8 @@ public class FeedbackEventRankedTest {
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.create(CaliperJsonldContext.V1P1_FEEDBACK.value());
-        id = "urn:uuid:a502e4fc-24c1-11e9-ab14-d663bd873d93";
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P2.value());
+        id = "urn:uuid:d3543a73-e307-4190-a755-5ce7b3187bc5";
         actor = Person.builder().id(BASE_IRI.concat("/users/554433")).build();
         creators = new ArrayList<CaliperAgent>();
         creators.add(actor);
@@ -88,53 +83,19 @@ public class FeedbackEventRankedTest {
         object = DigitalResource.builder()
             .id(SECTION_IRI.concat("/resources/1/syllabus.pdf"))
             .name("Course Syllabus")
-            //.creators(creators)
+            .creators(creators)
             .mediaType("application/pdf")
             .isPartOf(collection)
             .dateCreated(new DateTime(2018, 8, 2, 11, 32, 0, 0, DateTimeZone.UTC))
             .build();
 
-        List<String> itemLabels = Lists.newArrayList();
-        itemLabels.add("Strongly Disagree");
-        itemLabels.add("Disagree");
-        itemLabels.add("Agree");
-        itemLabels.add("Strongly Agree");
-
-        List<String> itemValues = Lists.newArrayList();
-        itemValues.add("-2");
-        itemValues.add("-1");
-        itemValues.add("1");
-        itemValues.add("2");
-
-        scale = LikertScale.builder()
-            .id(BASE_IRI.concat("/scale/2"))
-            .scalePoints(4)
-            .itemLabels(itemLabels)
-            .itemValues(itemValues)
-            .build();
-
-        question = RatingScaleQuestion.builder()
-            .id(BASE_IRI.concat("/question/2"))
-            .questionPosed("Do you agree with the opinion presented?")
-            .scale(scale)
-            .build();
-
-        ratingComment = Comment.builder()
-            .id(SECTION_IRI.concat("/assess/1/items/6/users/665544/responses/1/comment/1"))
-            .commenter(actor)
-            .commentedOn(object)
-            .value("I like what you did here but you need to improve on...")
-            .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
-            .build();
-
-        generated = Rating.builder()
-            .id(BASE_IRI.concat("/users/554433/rating/1"))
-            .rater(actor)
-            .rated(object)
-            .question(question)
-            .selection("1")
-            .ratingComment(ratingComment)
-            .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
+        generated = DigitalResource.builder()
+            .id(SECTION_IRI.concat("/resources/1/syllabus_copy.pdf"))
+            .name("Course Syllabus (copy)")
+            .creators(creators)
+            .mediaType("application/pdf")
+            .isPartOf(collection)
+            .dateCreated(new DateTime(2018, 11, 15, 10, 05, 0, 0, DateTimeZone.UTC))
             .build();
 
         edApp = SoftwareApplication.builder().id(BASE_IRI).coercedToId(true).build();
@@ -150,7 +111,7 @@ public class FeedbackEventRankedTest {
             .member(Person.builder().id(actor.getId()).coercedToId(true).build())
             .organization(CourseSection.builder().id(group.getId()).coercedToId(true).build())
             .status(Status.ACTIVE)
-            .role(Role.LEARNER)
+            .role(Role.INSTRUCTOR)
             .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
             .build();
 
@@ -160,7 +121,7 @@ public class FeedbackEventRankedTest {
             .build();
 
         // Build event
-        event = buildEvent(Action.RANKED);
+        event = buildEvent(Profile.RESOURCE_MANAGEMENT, Action.COPIED);
     }
 
     @Test
@@ -168,7 +129,7 @@ public class FeedbackEventRankedTest {
         ObjectMapper mapper = TestUtils.createCaliperObjectMapper();
         String json = mapper.writeValueAsString(event);
 
-        String fixture = jsonFixture("fixtures/v1p1/caliperEventFeedbackRanked.json");
+        String fixture = jsonFixture("fixtures/v1p2/caliperEventResourceManagementCopied.json");
         JSONAssert.assertEquals(fixture, json, JSONCompareMode.NON_EXTENSIBLE);
     }
 
@@ -182,19 +143,20 @@ public class FeedbackEventRankedTest {
      * @param action
      * @return event
      */
-    private FeedbackEvent buildEvent(Action action) {
-        return FeedbackEvent.builder()
-            .context(context)
-            .id(id)
-            .actor(actor)
-            .action(action)
-            .object(object)
-            .generated(generated)
-            .eventTime(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
-            .edApp(edApp)
-            .group(group)
-            .membership(membership)
-            .session(session)
-            .build();
+    private ResourceManagementEvent buildEvent(CaliperProfile profile, CaliperAction action) {
+        return ResourceManagementEvent.builder()
+                .context(context)
+                .profile(profile)
+                .id(id)
+                .actor(actor)
+                .action(action)
+                .object(object)
+                .generated(generated)
+                .eventTime(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
+                .edApp(edApp)
+                .group(group)
+                .membership(membership)
+                .session(session)
+                .build();
     }
 }
