+++ src/test/java/org/imsglobal/caliper/v1p2/events/ResourceManagementEventCreatedTest.java
@@ -16,25 +16,22 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.imsglobal.caliper.v1p1.events;
+package org.imsglobal.caliper.v1p2.events;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
+import org.imsglobal.caliper.actions.CaliperAction;
 import org.imsglobal.caliper.context.CaliperJsonldContext;
 import org.imsglobal.caliper.context.JsonldContext;
 import org.imsglobal.caliper.context.JsonldStringContext;
-import org.imsglobal.caliper.entities.CaliperEntity;
-import org.imsglobal.caliper.entities.agent.CourseSection;
-import org.imsglobal.caliper.entities.agent.Membership;
-import org.imsglobal.caliper.entities.agent.Person;
-import org.imsglobal.caliper.entities.agent.Role;
-import org.imsglobal.caliper.entities.agent.SoftwareApplication;
-import org.imsglobal.caliper.entities.agent.Status;
-import org.imsglobal.caliper.entities.search.Query;
-import org.imsglobal.caliper.entities.search.SearchResponse;
+import org.imsglobal.caliper.entities.agent.*;
+import org.imsglobal.caliper.entities.resource.DigitalResource;
+import org.imsglobal.caliper.entities.resource.DigitalResourceCollection;
 import org.imsglobal.caliper.entities.session.Session;
-import org.imsglobal.caliper.events.SearchEvent;
+import org.imsglobal.caliper.events.ResourceManagementEvent;
+import org.imsglobal.caliper.profiles.CaliperProfile;
+import org.imsglobal.caliper.profiles.Profile;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.After;
@@ -44,58 +41,57 @@ import org.junit.experimental.categories.Category;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static com.yammer.dropwizard.testing.JsonHelpers.jsonFixture;
 
 @Category(org.imsglobal.caliper.UnitTest.class)
-public class SearchEventSearchedTest {
+public class ResourceManagementEventCreatedTest {
     private JsonldContext context;
     private String id;
     private Person actor;
-    private Person creator;
-    private Query query;
-    private List<CaliperEntity> results;
-    private SoftwareApplication catalog;
-    private SoftwareApplication object;
-    private SearchResponse generated;
-    private SoftwareApplication edApp;
+    private CourseSection section;
+    private DigitalResource object;
+    private DigitalResourceCollection collection;
     private CourseSection group;
+    private List<CaliperAgent> creators;
     private Membership membership;
+    private ResourceManagementEvent event;
     private Session session;
-    private SearchEvent event;
+    private SoftwareApplication edApp;
 
     private static final String BASE_IRI = "https://example.edu";
-    private static final String BASE_CATALOG_IRI = "https://example.edu/catalog";
+    private static final String SECTION_IRI = BASE_IRI.concat("/terms/201801/courses/7/sections/1");
 
     @Before
     public void setUp() throws Exception {
-        context = JsonldStringContext.create(CaliperJsonldContext.V1P1_SEARCH.value());
-        id = "urn:uuid:cb3878ed-8240-4c6d-9fee-77221810f5e4";
+        context = JsonldStringContext.create(CaliperJsonldContext.V1P2.value());
+        id = "urn:uuid:0c81f804-62ee-4953-81c5-62d9579c4369";
         actor = Person.builder().id(BASE_IRI.concat("/users/554433")).build();
-        creator = Person.builder().id(BASE_IRI.concat("/users/554433")).coercedToId(true).build();
-        object = SoftwareApplication.builder().id(BASE_CATALOG_IRI).build();
+        creators = new ArrayList<CaliperAgent>();
+        creators.add(actor);
+        section = CourseSection.builder().id(SECTION_IRI).build();
+
+        collection = DigitalResourceCollection.builder()
+            .id(SECTION_IRI.concat("/resources/1"))
+            .name("Course Assets")
+            .isPartOf(section)
+            .build();
+
+        object = DigitalResource.builder()
+            .id(SECTION_IRI.concat("/resources/1/syllabus.pdf"))
+            .name("Course Syllabus")
+            .creators(creators)
+            .mediaType("application/pdf")
+            .isPartOf(collection)
+            .dateCreated(new DateTime(2018, 8, 2, 11, 32, 0, 0, DateTimeZone.UTC))
+            .build();
+
         edApp = SoftwareApplication.builder().id(BASE_IRI).coercedToId(true).build();
-        catalog = SoftwareApplication.builder().id(BASE_CATALOG_IRI).coercedToId(true).build();
-
-        query = Query.builder()
-            .id(BASE_IRI.concat("/users/554433/search?query=IMS%20AND%20%28Caliper%20OR%20Analytics%29"))
-            .creator(creator)
-            .searchTarget(catalog)
-            .searchTerms("IMS AND (Caliper OR Analytics)")
-            .dateCreated(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
-            .build();
-
-        generated = SearchResponse.builder()
-            .id(BASE_IRI.concat("/users/554433/response?query=IMS%20AND%20%28Caliper%20OR%20Analytics%29"))
-            .searchProvider(edApp)
-            .query(query)
-            .searchTarget(catalog)
-            .searchResultsItemCount(3)
-            .build();
 
         group = CourseSection.builder()
-            .id(BASE_IRI.concat("/terms/201801/courses/7/sections/1"))
+            .id(SECTION_IRI)
             .courseNumber("CPS 435-01")
             .academicSession("Fall 2018")
             .build();
@@ -105,7 +101,7 @@ public class SearchEventSearchedTest {
             .member(Person.builder().id(actor.getId()).coercedToId(true).build())
             .organization(CourseSection.builder().id(group.getId()).coercedToId(true).build())
             .status(Status.ACTIVE)
-            .role(Role.LEARNER)
+            .role(Role.INSTRUCTOR)
             .dateCreated(new DateTime(2018, 8, 1, 6, 0, 0, 0, DateTimeZone.UTC))
             .build();
 
@@ -115,22 +111,16 @@ public class SearchEventSearchedTest {
             .build();
 
         // Build event
-        event = buildEvent(Action.SEARCHED);
+        event = buildEvent(Profile.RESOURCE_MANAGEMENT, Action.CREATED);
     }
 
     @Test
     public void caliperEventSerializesToJSON() throws Exception {
         ObjectMapper mapper = TestUtils.createCaliperObjectMapper();
-
         String json = mapper.writeValueAsString(event);
 
-        String fixture = jsonFixture("fixtures/v1p1/caliperEventSearchSearched.json");
+        String fixture = jsonFixture("fixtures/v1p2/caliperEventResourceManagementCreated.json");
         JSONAssert.assertEquals(fixture, json, JSONCompareMode.NON_EXTENSIBLE);
-    }
-
-    @Test(expected=IllegalArgumentException.class)
-    public void viewEventRejectsNavigatedToAction() {
-        buildEvent(Action.NAVIGATED_TO);
     }
 
     @After
@@ -139,19 +129,19 @@ public class SearchEventSearchedTest {
     }
 
     /**
-     * Build View event
+     * Build Media event.
      * @param action
      * @return event
      */
-    private SearchEvent buildEvent(Action action) {
-        return SearchEvent.builder()
+    private ResourceManagementEvent buildEvent(CaliperProfile profile, CaliperAction action) {
+        return ResourceManagementEvent.builder()
                 .context(context)
+                .profile(profile)
                 .id(id)
                 .actor(actor)
                 .action(action)
                 .object(object)
                 .eventTime(new DateTime(2018, 11, 15, 10, 5, 0, 0, DateTimeZone.UTC))
-                .generated(generated)
                 .edApp(edApp)
                 .group(group)
                 .membership(membership)
